change offset from int to long in TupleState
Former-commit-id: c1c2a62337e3ea2ef5de18600df6911289d96577
Former-commit-id: fc41060297cac122a43b7ed0e7576790190bb591

+++ server/src/main/java/org/icgc/dcc/validation/cascading/TupleState.java
@@ -23,7 +23,7 @@ import com.google.common.collect.Lists;
 
 public class TupleState implements Serializable {
 
-  private final int offset;
+  private final long offset;
 
   private List<TupleError> errors;
 
@@ -67,7 +67,7 @@ public class TupleState implements Serializable {
     return structurallyValid;
   }
 
-  public int getOffset() {
+  public long getOffset() {
     return offset;
   }
